requirements.txt
Flask==2.3.0
Qrcode==7.4.2
pillow==10.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
pytest